requirements.txt
# Основные ML-зависимости
torch==2.0.1
transformers==4.30.0
sentence-transformers==2.2.2
accelerate==0.20.3

# Обработка документов
pdfminer.six==20221105
pytesseract==0.3.10
pillow==9.5.0
python-docx==0.8.11
openpyxl==3.1.2
pdf2image==1.16.3
PyPDF2==3.0.1

# Дополнительные
numpy==1.24.0
pandas==2.0.0
tqdm==4.65.0
loguru==0.7.0
flask==2.2.0
gunicorn==20.1.0
psycopg2-binary==2.9.6
requests==2.31.0